move pulsewatcher to compartment level
