+++ ledger-core/pulsewatcher/output/output.go
@@ -6,7 +6,7 @@
 package output
 
 import (
-	"github.com/insolar/assured-ledger/ledger-core/testutils/pulsewatcher/status"
+	"github.com/insolar/assured-ledger/ledger-core/pulsewatcher/status"
 )
 
 const (
