+++ ledger-core/pulsewatcher/output/text/emoji.go
